listen to create/delete user hooks to update the lookup server
